api/requirements.txt
firebase-functions
firebase-admin

# AI & Search
google-generativeai
pillow
requests
